+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Added new JSVM bindings:
   - `new Cookie({ ... })` constructor for creating `*http.Cookie` equivalent value.
   - `new SubscriptionMessage({ ... })` constructor for creating a custom realtime subscription payload.
+  - Soft-deprecated `$os.exec()` in favour of `$os.cmd()` to make it more clear that the call only prepares the command and doesn't execute it.
 
 
 ## v0.18.10